BJS2-68082: change name package, default value strength, agility, intellect moved to constants
